# express/src/server/swaggers.json
{
  "swagger": "2.0",
  "info": {
    "title": "zakaria api",
    "description": "the api of doom!!!",
    "version": "1.0.0"
  },
  "host": "api.zakariao.nl",
  "servers": [
    {
      "url": "api.zakariao.nl",
      "description": "Production api"
    },
    {
      "url": "jel.zakariao.nl",
      "description": "second api endpoint"
    }
  ],
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/api/pokea/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/poke/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  }
}